requirements.txt
fastapi
uvicorn
langchain
openai
python-dotenv
pydantic
streamlit
requests
